+++ polardbai-20251013/alibabacloud_polardbai20251013/models/_create_multimodal_search_task_shrink_request.py
@@ -10,6 +10,7 @@ class CreateMultimodalSearchTaskShrinkRequest(DaraModel):
         dbcluster_id: str = None,
         dataset_ids_shrink: str = None,
         embedding_model: str = None,
+        model_mode: str = None,
         query: str = None,
         search_model: str = None,
         top_k: int = None,
@@ -18,6 +19,7 @@ class CreateMultimodalSearchTaskShrinkRequest(DaraModel):
         self.dbcluster_id = dbcluster_id
         self.dataset_ids_shrink = dataset_ids_shrink
         self.embedding_model = embedding_model
+        self.model_mode = model_mode
         self.query = query
         self.search_model = search_model
         self.top_k = top_k
@@ -38,6 +40,9 @@ class CreateMultimodalSearchTaskShrinkRequest(DaraModel):
 
         if self.embedding_model is not None:
             result['EmbeddingModel'] = self.embedding_model
+
+        if self.model_mode is not None:
+            result['ModelMode'] = self.model_mode
 
         if self.query is not None:
             result['Query'] = self.query
@@ -60,6 +65,9 @@ class CreateMultimodalSearchTaskShrinkRequest(DaraModel):
 
         if m.get('EmbeddingModel') is not None:
             self.embedding_model = m.get('EmbeddingModel')
+
+        if m.get('ModelMode') is not None:
+            self.model_mode = m.get('ModelMode')
 
         if m.get('Query') is not None:
             self.query = m.get('Query')